requirements.txt
requests
rich
dnspython
email-validator
PyYAML
